Moves utils to separate folder
