+++ src/components/JarCard/utils/jarStepHandler.js
@@ -1,9 +1,9 @@
-import JarStepOne from './images/JarStep1.png';
-import JarStepTwo from './images/JarStep2.png';
-import JarStepThree from './images/JarStep3.png';
-import JarStepFour from './images/JarStep4.png';
-import JarStepFive from './images/JarStep5.png';
-import JarStepSix from './images/JarStep6.png';
+import JarStepOne from '../images/JarStep1.png';
+import JarStepTwo from '../images/JarStep2.png';
+import JarStepThree from '../images/JarStep3.png';
+import JarStepFour from '../images/JarStep4.png';
+import JarStepFive from '../images/JarStep5.png';
+import JarStepSix from '../images/JarStep6.png';
 
 const jarStepHandler = (value, goal) => {
   const JarImagesArr = [
